requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
SQLAlchemy==1.4.46
psycopg2-binary==2.9.6
alembic==1.9.2
passlib==1.7.4
python-jose==3.3.0